Fix refresh when double-cell characters end a line
Double cell characters that end a line are not properly displayed
when there are split regions. Takeshi Banse detected the problem and
a fix for it in Savannag bug #29106. This is a better fix, as it
will catch some more cases of the same problem. But this particular
fix is usable only with the changes made for better unicode support
in unicode++ branch.